refactor: move task item UI to a separate component
