+++ src/components/TasksList/TasksList.module.scss
@@ -6,57 +6,3 @@
     align-items: center;
     justify-content: center;
 }
-
-.item {
-    width: 100%;
-    margin-bottom: 15px;
-    &-inner {
-        width: 100%;
-        padding: 15px;
-        margin-bottom: 0px;
-    }
-}
-
-.title {
-    font-size: 24px;
-    font-weight: 600;
-    margin-bottom: 10px;
-    text-align: left;
-}
-
-.description {
-    font-size: 16px;
-    font-weight: 500;
-    text-align: left;
-}
-
-.priority {
-    font-weight: bold;
-    text-align: end;
-    margin-bottom: 10px;
-
-    display: flex;
-    gap: 10px;
-    align-items: center;
-    justify-content: flex-end;
-}
-
-.actions {
-    display: flex;
-    gap: 15px;
-}
-
-.bullet {
-    width: 30px;
-    height: 30px;
-    border-radius: 100%;
-    &-high {
-        background-color: red;
-    }
-    &-mid {
-        background-color: orange;
-    }
-    &-low {
-        background-color: green;
-    }
-}
